the first pass at an implementation of keys
This is dirty as fuck but is the first implementation which passes
tests. I may need to write more tests, and I also have to handle the
case of async unmount, and I really am itching to refactor this, but at
least I now know its possible to have persistent/keyed nodes which move
around!