revert checking props because it broke TodoMVC
Unlike other frameworks, we can’t assume that old and new props being
referentially equal means the underlying node shouldn’t be updated, so
we just update the prop to match what’s been given. Maybe we can
consider this optimization in the future, but it doesn’t seem cranky to
do so.