Implement hydration for the sync case 3.
An attempt at implementing hydration. The problem with this approach is that it
does not correctly handle async siblings. Hydration has to connect real DOM
nodes to virtual elements, but the relationship between these is many to many.
For instance, a component element can have zero, one, or many DOM nodes as
children. Combined with the fact that subtrees can be async, this means that we
cannot know which virtual siblings correspond to which hydrated nodes until
each virtual element’s previous sibling has settled.

Potential solutions:
1. For hydration, run updating of siblings in sequence, rather than in parallel.
  - Honestly this makes sense to me as a limitation of hydration. Long-running async trees for
    hydration does not seem wise.
2. Run hydration on an off-screen DOM fragment and use something like morphdom to patch the active document.
  - Still need to figure out how to connect event handlers and retainers to DOM nodes.
3. Disallow hydration with async components.
  - Nahhhh