a framework for frameworks pt. 1
I have learned not to pray to pagan gods in commit messages, not even in jest.

Forgive me, Oh Lord, for that was a bad joke and probably wrong to do. Please
let me not re-invent the wheel, let the framework be nice and easy to use, let it not take much longer than a week.

In Jesus’s name we pray in commit messages Amen.