Allow overriding xmlns
A foreignObject within an svg tag should not lose its SVG_NAMESPACE,
because it is still an SVG element.

In addition, children of a foreignObject element should be allowed to
specify their xmlns, since it is likely they will be some other NS than
the foreignObject, e.g. xhtml.

This PR allows elements to specify their xmlns, when needed.